refactor gmix_ndim/jacobian to sub-packages
