+++ ngmix/gmix_ndim/__init__.py
@@ -1,0 +1,6 @@
+# flake8: noqa
+
+from . import gmix_ndim
+from .gmix_ndim import *
+
+from . import gmix_ndim_nb
