+++ ngmix/jacobian/__init__.py
@@ -1,0 +1,5 @@
+# flake8: noqa
+
+from . import jacobian
+from .jacobian import *
+from . import jacobian_nb
